backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
selenium==4.12.0
webdriver-manager==4.0.0
websockets==11.0.3
pydantic==2.3.0
python-dotenv==1.0.0